+++ packages/nodes-base/package.json
@@ -99,7 +99,7 @@
       "dist/credentials/JiraSoftwareCloudApi.credentials.js",
       "dist/credentials/JiraSoftwareServerApi.credentials.js",
       "dist/credentials/JotFormApi.credentials.js",
-      "dist/credentials/ApacheKafka.credentials.js",
+      "dist/credentials/Kafka.credentials.js",
       "dist/credentials/KeapOAuth2Api.credentials.js",
       "dist/credentials/LinkedInOAuth2Api.credentials.js",
       "dist/credentials/MailchimpApi.credentials.js",
@@ -292,7 +292,7 @@
       "dist/nodes/Jira/Jira.node.js",
       "dist/nodes/Jira/JiraTrigger.node.js",
       "dist/nodes/JotForm/JotFormTrigger.node.js",
-      "dist/nodes/ApacheKafka/ApacheKafka.node.js",
+      "dist/nodes/Kafka/Kafka.node.js",
       "dist/nodes/Keap/Keap.node.js",
       "dist/nodes/Keap/KeapTrigger.node.js",
       "dist/nodes/LinkedIn/LinkedIn.node.js",
